Develop story summary based on notes and initial concept
Let's think about where we want to take my new short story called "The Check". Here is an early sketch of the idea from 3 years ago: https://shaisachs.com/2022/06/29/the-check-first-thoughts.html

Here are some rough notes I sketched out more recently:

---

Based in the diner
Uses LLMs to match people up for dates

Something goes horribly wrong because of a problem in the training data
How do we figure out something goes horribly wrong? How to make it suitably dramatic in obvious ways for a short story?

How do we figure out it's the training data?

Maybe something happens on a particular day Maybe this story should be about alienation and more personal - not about something bad happening to a lot of people because of bad data, but something bad happening to a single person because of alienated data?
---

Based on the above, why don't we try to write out a basic summary of what the plot should be, and who the characters are?