Target characters from earlier stories and cast about for a good victim
Well, the characters should be Frankie (from The Menu) and Augusta (from The Recipe). I'm also not sure I agree with your mapping of this story into kishotenketsu structure.

As a practical matter I also find that it's difficult, as a storytelling matter, to dramatize the problem with the data, and yet more difficult to dramatize a problem with alienation. I'm not sure who should be the "victim" of this machine - its inventor? Or its customers? What do you think?