Switch roles. Abandon kishotenketsu. Explore inventor-as-victim.
No, Frankie should be the owner and Augusta the inventor. That is more continuous with the previous stories.

I do not think we should follow a kishotenketsu format for this story. I tried that in The Game and I enjoyed it, but that does not seem like the right fit here.

I don't want to explore "diner-optimization" scenarios, since I did that in The Menu. Anyway I don't find it very interesting.

I like the idea of scenario A. Maybe we can have some kind of matchmaking chatbot that is operational in the diner (but how?) but it gradually loses context on the matches it produces (and how do we demonstrate that in show-don't-tell fashion, from a storytelling perspective). What would that look like?