Frankie and Augusta stay friends. Focus on indexical data.
We are definitely getting closer. I don't actually like the idea of a confrontation between Augusta and Frankie, I think they should remain friends on good terms throughout. More importantly the resolution seems kind of boring and unconvincing. The collapse is also not that dramatic.

More problematic still is the problem of "indexical" data. I think fundamentally what I want to draw out is the problem of context that can't be materialized into text because it is somehow too ineffable. Context like "Martin died a decade ago" is easy to write down, even at scale. But I'm not sure of what kind of ineffable context might be. Perhaps it is something shared between Frankie and Augusta that they do not even realize they both understand. What might that be?