Explore check-that-isn't-a-check. Kill off Johnny.
I like the idea of a check that isn't a check. But I don't think of Augusta as a long-standing partner to Frankie. She is a cousin or perhaps a sister who is there for a visit. One of the things I am thinking, but am not sure of, is that perhaps Frankie's husband (Johhny, also from The Menu) has died and Augusta is visiting to keep her cousin company. Something like that. Johnny was the "front of the house" in The Menu. Perhaps the "ineffable context" has something to do with the memory of him?