Pivot - not about data out of context but AI out of context. Revisit kishotenketsu.1
I've given this story some thought and I want to try a pivot.

The point of the story is still that it's about data taken out of context, but not in a way that makes the AI into some kind of enemy or boogey man. The point is more about using AI in inappropriate ways. In part this change reflects my view that data out of context is not an intrinsic evil - after all, information wants to be free, and so removing it from its source context is inevitable. Rather the evil is in using tools in destructive ways.

I like the idea that this story is set in the same diner setting from The Menu. Johnny has died and left behind boxes and boxes of old receipts, as well as a legacy of being the town matchmaker. But Johnny's death is not made explicit. It is at best hinted at, and perhaps not even that.

I also want to return to the kishotenketsu format, in a way that brings harmony to this story as well as to this series as a whole.

Here is how I think the story breaks down.

1. Introduction
We begin with the AI matchmaker already having been invented. I imagine an opening scene in which a customer looks at his check, notices a somewhat quirky suggested match printed on the last line, and has a bemused conversation with Frankie about it.

2. Development
We go into the backstory of how the matchmaker was created. Augusta, from The Recipe, had come to help Frankie clear out the diner, and discovered Johnny's old receipts. She gets the idea to digitize them and leverage them for matchmaking, though I'm not yet sure how. The matchmaker becomes a hit and customers travel from far and wide to try it out.

3. Complication
We turn to a new setting, a corporate environment of some kind, in which hiring and firing decisions are made in inscrutable ways. There is a kind of constant chaos, and employees are left baffled about their future and how they are supposed to get anything done. We zero in on one employee - who is Mathi from The Recipe. All of his teammates have just been fired and he has been assigned to work with someone he's never met, who turns out to be the unnamed narrator of The Game. (We need to give her a name, though I'm not sure what it should be. Since Mathi is supposed to be a kind of arch reference to Alan Turing, perhaps she should have a name reminiscent of Alonzo Church.) I am also not sure how to make it plain that she is the same person as the narrator of The Game. The two characters start talking, and it turns out they work together very nicely - they even have the same taste in lunch orders.

4. Resolution
We find ourselves at some unspecified point in the future. Mathi and his coworker have left their job and are working together on some kind of do-gooder passion project; they in fact use the diner as their favorite working spot. We learn that the corporation where they worked was using their lunch receipts (which were submitted for reimbursement) to determine who they should be paired with, and making hiring and firing decisions on the basis of these pairings. It is implied, though perhaps not specified, that the corporation eventually went kaput; meanwhile Mathi and his coworker are happily engaged in something they enjoy. Not only that, but Augusta and Frankie have learned that their matchmaker has this kind of potential, and they use it to bring people together for similar projects. They are so successful that they open up a chain of diner/coworking spaces and bring many people together in this way. I am tempted to give the diner a name drawn from Micah - The Vine and Fig Tree - though perhaps that is a bit corny. I'm not really sure how to wrap up the story.

One problem I am thinking about is, how do I incorporate techno-magical realism into this plot? As I've described it so far, the AI is kind of a constant throughout the story.