Expand part 3
All right, now we're focused only on part 3. We want it to be about double its current length, maybe even a little longer. We want to really feel the contrast between the whimsical, magical atmosphere of part 2 and the alienating, mis-aligned, incongruous feeling of part 3. We don't really want to be too explicit about how the matchmaker is being misused. At the same time we want to feel the connection between Mathi and Alma, it should be understated but obvious.