Expand part 4 to be more fun. Explore Micah.
Ok now we're focused on part 4. It's really short, only 700 words. I'd like it to be closer to 1500 words. Partly because it's the most fun part. Also I want to expand a little on the idea of the "network" though I don't really like that exact word.

I am also interested in the phrase "under their own vine and fig tree", which appears in Micah and elsewhere. I'd be interested to infuse some of that sensibility into this part, though it should not be too overt.